requirements.txt
pandas==1.3.5
numpy==1.21.6
matplotlib==3.5.1
scikit-learn==1.0.2
imgaug==0.4.0
opencv-python-headless==4.5.5.64
tensorflow==2.8.0
keras==2.8.0
eventlet==0.33.0
Flask==2.0.2
socketio==0.2.1
Pillow==8.4.0

